Fixed bug where halt mode is not stopped when an interrupt is triggered
This is because of
```
graphics.Update(cycles);
interrupts.Update();
```
If an interrupt was triggered in `graphics` for example, `Cpu::ServiceInterrupt`
will be called from `Interrupts.cs` but the halt mode will not be cleared until
another interrupt is requested.

The cause for this is in the line
```
if (haltMode != HaltMode.NotHalted && haltMode != HaltMode.HaltBug && interrupts.HasPendingInterrupts)
```
Since the interrupt is done being processed, `HasPendingInterrupts`
will always return `False`.